Add null resolve tests + simplify test classes structure.
