+++ src/test/java/io/opentracing/contrib/tracerresolver/Mocks.java
@@ -1,0 +1,28 @@
+package io.opentracing.contrib.tracerresolver;
+
+import io.opentracing.Tracer;
+import io.opentracing.mock.MockTracer;
+
+public final class Mocks {
+
+    public static class FallbackTracer extends MockTracer {
+    }
+
+    public static class ResolvedTracer extends MockTracer {
+    }
+
+    public static class MockTracerResolver extends TracerResolver {
+        @Override
+        protected Tracer resolve() {
+            return new ResolvedTracer();
+        }
+    }
+
+    public static class NullTracerResolver extends TracerResolver {
+        @Override
+        protected Tracer resolve() {
+            return null;
+        }
+    }
+
+}
