--- src/test/java/io/opentracing/contrib/tracerresolver/ServiceFileUtil.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2017 The OpenTracing Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.opentracing.contrib.tracerresolver;
-
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.PrintWriter;
-
-final class ServiceFileUtil {
-
-    static void writeServiceFile(File serviceFile, String content) throws IOException {
-        serviceFile.getParentFile().mkdirs();
-        if (serviceFile.isFile()) serviceFile.delete();
-        PrintWriter writer = new PrintWriter(new FileWriter(serviceFile));
-        try {
-            writer.println(content);
-        } finally {
-            writer.close();
-        }
-    }
-
-}
