--- src/test/java/io/opentracing/contrib/tracerresolver/TracerResolverFallbackTest.java
@@ -1,60 +1,0 @@
-/*
- * Copyright 2017 The OpenTracing Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.opentracing.contrib.tracerresolver;
-
-import io.opentracing.Tracer;
-import io.opentracing.mock.MockTracer;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-
-import java.io.File;
-import java.io.IOException;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.instanceOf;
-import static org.hamcrest.Matchers.is;
-
-public class TracerResolverFallbackTest {
-    private static final File TRACER_SERVICE_FILE =
-            new File("target/test-classes/META-INF/services/" + Tracer.class.getName());
-
-    @Before
-    public void prepareClasspath() throws IOException {
-        ServiceFileUtil.writeServiceFile(TRACER_SERVICE_FILE, TestTracer.class.getName());
-    }
-
-    @After
-    public void restoreClasspath() throws IOException {
-        TRACER_SERVICE_FILE.delete();
-    }
-
-    @Before
-    @After
-    public void resetTracerResolver() {
-        TracerResolver.reload();
-    }
-
-    @Test
-    public void testResolveFallback() {
-        assertThat(TracerResolver.resolveTracer(), is(instanceOf(TestTracer.class)));
-    }
-
-    public static class TestTracer extends MockTracer {
-
-    }
-
-}
